+++ wom/services/groups.py
@@ -34,8 +34,8 @@ from . import BaseService
 __all__ = ("GroupService",)
 
 
-ValueT = t.TypeVar("ValueT")
-ResultT = result.Result[ValueT, models.HttpErrorResponse]
+T = t.TypeVar("T")
+ResultT = result.Result[T, models.HttpErrorResponse]
 
 
 class GroupService(BaseService):
@@ -71,7 +71,9 @@ class GroupService(BaseService):
 
         Args:
             name: The group name to search for.
+
             limit: The pagination limit.
+
             offset: The pagination offset.
 
         Returns:
@@ -92,12 +94,12 @@ class GroupService(BaseService):
         """
         params = self._generate_map(name=name, limit=limit, offset=offset)
         route = routes.SEARCH_GROUPS.compile().with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_group(p) for p in data])
+        return self._ok(data, t.List[models.Group])
 
     async def get_details(self, id: int) -> ResultT[models.GroupDetail]:
         """Gets the details for the given group id.
@@ -121,12 +123,12 @@ class GroupService(BaseService):
             ```
         """
         route = routes.GROUP_DETAILS.compile(id)
-        data = await self._http.fetch(route, self._dict)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok(self._serializer.deserialize_group_details(data))
+        return self._ok(data, models.GroupDetail)
 
     async def create_group(
         self,
@@ -135,7 +137,7 @@ class GroupService(BaseService):
         clan_chat: t.Optional[str] = None,
         description: t.Optional[str] = None,
         homeworld: t.Optional[int] = None,
-    ) -> ResultT[models.GroupDetail]:
+    ) -> ResultT[models.CreatedGroupDetail]:
         """Creates a new group.
 
         Args:
@@ -154,13 +156,13 @@ class GroupService(BaseService):
                 `None`.
 
         Returns:
-            A [`Result`][wom.Result] containing the group details.
+            A [`Result`][wom.Result] containing the created group details.
 
         !!! note
 
-             A mixture of strings and GroupMemberFragments can be passed for
-             members. If a string is passed, no role will be added for that
-             member.
+            A mixture of strings and GroupMemberFragments can be passed for
+            members. If a string is passed, no role will be added for that
+            member.
 
         ??? example
 
@@ -190,15 +192,12 @@ class GroupService(BaseService):
         )
 
         route = routes.CREATE_GROUP.compile()
-        data = await self._http.fetch(route, self._dict, payload=payload)
+        data = await self._http.fetch(route, payload=payload)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        # Verification code is only present on new group creations
-        group = self._serializer.deserialize_group_details(data["group"])
-        group.verification_code = data["verificationCode"]
-        return result.Ok(group)
+        return self._ok(data, models.CreatedGroupDetail)
 
     async def edit_group(
         self,
@@ -285,12 +284,12 @@ class GroupService(BaseService):
         )
 
         route = routes.EDIT_GROUP.compile(id)
-        data = await self._http.fetch(route, self._dict, payload=payload)
+        data = await self._http.fetch(route, payload=payload)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok(self._serializer.deserialize_group_details(data))
+        return self._ok(data, models.GroupDetail)
 
     async def delete_group(
         self, id: int, verification_code: str
@@ -322,14 +321,10 @@ class GroupService(BaseService):
             await client.groups.delete_group(123, "111-111-111")
             ```
         """
-        payload = self._generate_map(verificationCode=verification_code)
         route = routes.DELETE_GROUP.compile(id)
-        data = await self._http.fetch(route, models.HttpErrorResponse, payload=payload)
-
-        if not data.message.startswith("Success"):
-            return result.Err(data)
-
-        return result.Ok(models.HttpSuccessResponse(data.status, data.message))
+        payload = self._generate_map(verificationCode=verification_code)
+        data = await self._http.fetch(route, payload=payload, allow_http_success=True)
+        return self._success_or_err(data)
 
     async def add_members(
         self, id: int, verification_code: str, *members: t.Union[str, models.GroupMemberFragment]
@@ -379,12 +374,8 @@ class GroupService(BaseService):
         )
 
         route = routes.ADD_MEMBERS.compile(id)
-        data = await self._http.fetch(route, models.HttpErrorResponse, payload=payload)
-
-        if not data.message.startswith("Success"):
-            return result.Err(data)
-
-        return result.Ok(models.HttpSuccessResponse(data.status, data.message))
+        data = await self._http.fetch(route, payload=payload, allow_http_success=True)
+        return self._success_or_err(data)
 
     async def remove_members(
         self, id: int, verification_code: str, *members: str
@@ -419,15 +410,10 @@ class GroupService(BaseService):
             )
             ```
         """
-        payload = self._generate_map(verificationCode=verification_code, members=members)
-
         route = routes.REMOVE_MEMBERS.compile(id)
-        data = await self._http.fetch(route, models.HttpErrorResponse, payload=payload)
-
-        if not data.message.startswith("Success"):
-            return result.Err(data)
-
-        return result.Ok(models.HttpSuccessResponse(data.status, data.message))
+        payload = self._generate_map(verificationCode=verification_code, members=members)
+        data = await self._http.fetch(route, payload=payload, allow_http_success=True)
+        return self._success_or_err(data)
 
     async def change_member_role(
         self, id: int, verification_code: str, username: str, role: models.GroupRole
@@ -469,17 +455,26 @@ class GroupService(BaseService):
         )
 
         route = routes.CHANGE_MEMBER_ROLE.compile(id)
-        data = await self._http.fetch(route, self._dict, payload=payload)
+        data = await self._http.fetch(route, payload=payload)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok(self._serializer.deserialize_group_membership(data))
+        return self._ok(data, models.GroupMembership)
 
     async def update_outdated_members(
         self, id: int, verification_code: str
     ) -> ResultT[models.HttpSuccessResponse]:
         """Attempts to update all outdated group members.
+
+        Args:
+            id: The ID of the group.
+
+            verification_code: The verification code for the group.
+
+        Returns:
+            A [`Result`][wom.Result] containing the success response
+                message.
 
         !!! info
 
@@ -512,24 +507,11 @@ class GroupService(BaseService):
                 123, "111-111-111"
             )
             ```
-
-        Args:
-            id: The ID of the group.
-
-            verification_code: The verification code for the group.
-
-        Returns:
-            A [`Result`][wom.Result] containing the success response
-                message.
         """
-        payload = self._generate_map(verificationCode=verification_code)
         route = routes.UPDATE_OUTDATED_MEMBERS.compile(id)
-        data = await self._http.fetch(route, models.HttpErrorResponse, payload=payload)
-
-        if "players are being updated" in data.message:
-            return result.Ok(models.HttpSuccessResponse(data.status, data.message))
-
-        return result.Err(data)
+        payload = self._generate_map(verificationCode=verification_code)
+        data = await self._http.fetch(route, payload=payload, allow_http_success=True)
+        return self._success_or_err(data, predicate=lambda m: "players are being updated" in m)
 
     async def get_competitions(
         self, id: int, *, limit: t.Optional[int] = None, offset: t.Optional[int] = None
@@ -562,12 +544,12 @@ class GroupService(BaseService):
         """
         params = self._generate_map(limit=limit, offset=offset)
         route = routes.GROUP_COMPETITIONS.compile(id).with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_competition(c) for c in data])
+        return self._ok(data, t.List[models.Competition])
 
     async def get_gains(
         self,
@@ -607,7 +589,7 @@ class GroupService(BaseService):
 
         !!! info
 
-            You can pass either (`period`) or (`start_date` +
+            You must pass one of (`period`) or (`start_date` +
             `end_date`), but not both.
 
         ??? example
@@ -620,7 +602,7 @@ class GroupService(BaseService):
             await client.start()
 
             await client.groups.get_gains(
-                123, wom.Bosses.Zulrah, limit=10
+                123, wom.Metric.Zulrah, limit=10
             )
             ```
         """
@@ -634,12 +616,12 @@ class GroupService(BaseService):
         )
 
         route = routes.GROUP_GAINS.compile(id).with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_group_member_gains(d) for d in data])
+        return self._ok(data, t.List[models.GroupMemberGains])
 
     async def get_achievements(
         self,
@@ -675,12 +657,12 @@ class GroupService(BaseService):
         """
         params = self._generate_map(limit=limit, offset=offset)
         route = routes.GROUP_ACHIEVEMENTS.compile(id).with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_achievement(a) for a in data])
+        return self._ok(data, t.List[models.Achievement])
 
     async def get_records(
         self,
@@ -719,7 +701,7 @@ class GroupService(BaseService):
             await client.start()
 
             await client.groups.get_records(
-                123, wom.Bosses.Zulrah, wom.Period.Day, limit=3
+                123, wom.Metric.Zulrah, wom.Period.Day, limit=3
             )
             ```
         """
@@ -731,12 +713,12 @@ class GroupService(BaseService):
         )
 
         route = routes.GROUP_RECORDS.compile(id).with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_record_leaderboard_entry(a) for a in data])
+        return self._ok(data, t.List[models.RecordLeaderboardEntry])
 
     async def get_hiscores(
         self,
@@ -772,18 +754,19 @@ class GroupService(BaseService):
             await client.start()
 
             await client.groups.get_hiscores(
-                123, wom.Skills.Runecrafting, limit=10
+                123, wom.Metric.Runecrafting, limit=10
             )
             ```
         """
         params = self._generate_map(limit=limit, offset=offset, metric=metric.value)
         route = routes.GROUP_HISCORES.compile(id).with_params(params)
-        data = await self._http.fetch(route, self._list)
+        data = await self._http.fetch(route)
 
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return result.Ok([self._serializer.deserialize_group_hiscores_entry(h) for h in data])
+        # TODO: hiscores entries are broken as hell
+        return self._ok(data, t.List[models.GroupHiscoresEntry])
 
     async def get_name_changes(
         self, id: int, *, limit: t.Optional[int] = None, offset: t.Optional[int] = None