+++ wom/services/players.py
@@ -33,8 +33,8 @@ from . import BaseService
 
 __all__ = ("PlayerService",)
 
-ValueT = t.TypeVar("ValueT")
-ResultT = result.Result[ValueT, models.HttpErrorResponse]
+T = t.TypeVar("T")
+ResultT = result.Result[T, models.HttpErrorResponse]
 
 
 class PlayerService(BaseService):
@@ -80,7 +80,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.Player])
+        return self._ok(data, t.List[models.Player])
 
     async def update_player(self, username: str) -> ResultT[models.PlayerDetail]:
         """Updates the given player.
@@ -110,7 +110,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, models.PlayerDetail)
+        return self._ok(data, models.PlayerDetail)
 
     async def assert_player_type(self, username: str) -> ResultT[models.AssertPlayerType]:
         """Asserts, and fixes, a players type.
@@ -140,7 +140,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, models.AssertPlayerType)
+        return self._ok(data, models.AssertPlayerType)
 
     async def get_details(self, username: str) -> ResultT[models.PlayerDetail]:
         """Gets the details for a given player.
@@ -169,7 +169,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, models.PlayerDetail)
+        return self._ok(data, models.PlayerDetail)
 
     async def get_details_by_id(self, player_id: int) -> ResultT[models.PlayerDetail]:
         """Gets the details for a given player id.
@@ -198,7 +198,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, models.PlayerDetail)
+        return self._ok(data, models.PlayerDetail)
 
     async def get_achievements(self, username: str) -> ResultT[t.List[models.Achievement]]:
         """Gets the achievements for a given player.
@@ -228,7 +228,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.Achievement])
+        return self._ok(data, t.List[models.Achievement])
 
     async def get_achievement_progress(
         self, username: str
@@ -260,7 +260,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.PlayerAchievementProgress])
+        return self._ok(data, t.List[models.PlayerAchievementProgress])
 
     async def get_competition_participations(
         self,
@@ -299,7 +299,9 @@ class PlayerService(BaseService):
 
             await client.start()
 
-            result = await client.players.get_achievement_progress("Jonxslays")
+            result = await client.players.get_competition_participations(
+                "Jonxslays", limit=3
+            )
             ```
         """
         params = self._generate_map(
@@ -314,7 +316,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.PlayerParticipation])
+        return self._ok(data, t.List[models.PlayerParticipation])
 
     async def get_competition_standings(
         self,
@@ -353,7 +355,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.PlayerCompetitionStanding])
+        return self._ok(data, t.List[models.PlayerCompetitionStanding])
 
     async def get_group_memberships(
         self, username: str, *, limit: t.Optional[int] = None, offset: t.Optional[int] = None
@@ -395,7 +397,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.PlayerMembership])
+        return self._ok(data, t.List[models.PlayerMembership])
 
     async def get_gains(
         self,
@@ -425,7 +427,7 @@ class PlayerService(BaseService):
 
         !!! info
 
-            You can pass either (`period`) or (`start_date` +
+            You must pass one of (`period`) or (`start_date` +
             `end_date`), but not both.
 
         ??? example
@@ -454,7 +456,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, models.PlayerGains)
+        return self._ok(data, models.PlayerGains)
 
     async def get_records(
         self,
@@ -489,7 +491,7 @@ class PlayerService(BaseService):
             await client.start()
 
             result = await client.players.get_records(
-                "Jonxslays", period=wom.Period.Day, metric=wom.Skills.Attack
+                "Jonxslays", period=wom.Period.Day, metric=wom.Metric.Attack
             )
             ```
         """
@@ -503,7 +505,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.Record])
+        return self._ok(data, t.List[models.Record])
 
     async def get_snapshots(
         self,
@@ -533,7 +535,7 @@ class PlayerService(BaseService):
 
         !!! info
 
-            You can pass either (`period`) or (`start_date` +
+            You must pass one of (`period`) or (`start_date` +
             `end_date`), but not both.
 
         ??? example
@@ -562,7 +564,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.Snapshot])
+        return self._ok(data, t.List[models.Snapshot])
 
     async def get_name_changes(self, username: str) -> ResultT[t.List[models.NameChange]]:
         """Gets the name changes for the player.
@@ -591,7 +593,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.NameChange])
+        return self._ok(data, t.List[models.NameChange])
 
     async def get_snapshots_timeline(
         self,
@@ -625,7 +627,7 @@ class PlayerService(BaseService):
 
         !!! info
 
-            You can pass either (`period`) or (`start_date` +
+            You must pass one of (`period`) or (`start_date` +
             `end_date`), but not both.
 
         ??? example
@@ -655,7 +657,7 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.SnapshotTimelineEntry])
+        return self._ok(data, t.List[models.SnapshotTimelineEntry])
 
     async def get_archives(
         self,
@@ -687,4 +689,4 @@ class PlayerService(BaseService):
         if isinstance(data, models.HttpErrorResponse):
             return result.Err(data)
 
-        return self.ok(data, t.List[models.PlayerArchive])
+        return self._ok(data, t.List[models.PlayerArchive])